Séparation de la gestion des étapes dans un autre controlleur
